Updated staking page instructions
Hey @laboon - I went to stake on Kusama today and couldn't find any instructions on how to nominate. My path was Google>"How to stake on Kusama">I found the "Staking" page on this guide, but what I really needed was the "Nominating" page. The whole industry uses the word staking instead of nominating, so this could cause some issues. I was only able to nominate after I found a medium article buried on the Kusama forum. I'm just updating this page to add a couple call-outs to the Nominating instructions in case other people do the same thing I did.